Fall back on scipy's SVD if numpy's throws ValueError
numpy.linalg.svd sometimes throws a ValueError even if the matrix
doesn't have any inf's or nan's. We therefore now try to catch the
ValuError, and retry with scipy.linalg.svd with the
lapack_driver='gesvd' argument. If a ValueError still occurs, let the
user know if the matrix has any inf's or nan's and re-raise the
Exception.